// frontend/src/api/generated/model/cronRunEntry.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * Mission Control API
 * OpenAPI spec version: 0.1.0
 */
import type { CronRunEntryRaw } from "./cronRunEntryRaw";

export interface CronRunEntry {
  id?: string | null;
  started_at?: string | null;
  finished_at?: string | null;
  status?: string | null;
  raw?: CronRunEntryRaw;
}