// frontend/src/api/generated/model/cronUpdateRequest.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * Mission Control API
 * OpenAPI spec version: 0.1.0
 */

export interface CronUpdateRequest {
  schedule?: string | null;
  target_agent?: string | null;
  message?: string | null;
  name?: string | null;
  enabled?: boolean | null;
}